HGR on a real display: VGA mode 13h for DOS, half-blocks natively
Graphics previously wrote correct bytes into page memory and nothing showed
them. Now they are displayed, and the colour model is right.

Colour. HCOLOR does not select a colour so much as a bit pattern: green only
lights odd columns and violet only even ones, so a single green dot at an
even x draws nothing at all. Colour is a property of position, not of the
bit - the pixel clock ran at the colour subcarrier frequency, so where a dot
sat decided what came out. A lit pixel reads white if either neighbour is
lit, and a lone dot fills a whole colour cycle, which is why a line plotted
on every other column looks solid rather than dotted. That widening happens
in the rasterizer, so page memory stays honest and PEEK still sees one bit.

The visible consequence: a one-pixel vertical line plotted in white comes out
violet on an even column and green on an odd one. That is the hardware.

Display. display_dos.c drives VGA mode 13h - 320x200 fits hi-res at 280x192
with a border, and lo-res at 40x48 scales to exactly 8x4 per cell. Plotted
pixels go straight to video memory as they happen; a POKE or a screen clear
asks for a full repaint. display_term.c draws the same pages with terminal
half-blocks so the native build can show a picture too.

Two bugs found and fixed along the way: video memory was not being cleared on
entering graphics, so stray bytes showed up as stripes in the BIOS default
palette; and 320 * 200 overflows a signed 16-bit int, so the clear size had to
be built from unsigned operands.

Verified: the DOS build runs HGRDEMO.BAS to completion under DOSBox with
ERRNUM 0 and correct page memory, and drew the border, sine and colour bars to
VGA. Because the emulator throttles when its pane is hidden, the risky logic
is tested on the host instead: all 7680 hi-res and 960 lo-res addresses invert
exactly, layout holes are rejected, and every partial repaint span agrees with
a full-row render.

Also adds -r to run a program without waiting for RUN.

+++ src/a2mem.h
@@ -57,14 +57,17 @@ a2addr a2_word(a2addr a);              /* little-endian 16-bit read */
 void   a2_setword(a2addr a, a2addr v);
 
 /* --- program ------------------------------------------------------------ */
+/* Line numbers are long throughout. Applesoft allows up to 63999, which does
+ * not fit in the 16-bit int of the DOS build -- LIST's upper bound of 65535
+ * silently became -1 and listed nothing until this was widened. */
 a2addr a2_prog_first(void);                    /* 0 when empty */
 a2addr a2_prog_next(a2addr line);              /* 0 at end */
-int    a2_prog_lineno(a2addr line);
+long   a2_prog_lineno(a2addr line);
 const unsigned char *a2_prog_tokens(a2addr line);
-a2addr a2_prog_find(int lineno);               /* 0 if absent */
-a2addr a2_prog_find_ge(int lineno);            /* first line >= lineno */
-int    a2_prog_insert(int lineno, const unsigned char *toks, int len);
-void   a2_prog_delete(int lineno);
+a2addr a2_prog_find(long lineno);              /* 0 if absent */
+a2addr a2_prog_find_ge(long lineno);           /* first line >= lineno */
+int    a2_prog_insert(long lineno, const unsigned char *toks, int len);
+void   a2_prog_delete(long lineno);
 
 /* --- variables ---------------------------------------------------------- */
 #define VT_REAL 0
